backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
langchain==0.0.350
langchain-community==0.0.10
openai==1.3.7
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
PyPDF2==3.0.1
chromadb==0.4.18
eventlet
PyMuPDF==1.23.8
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
sentence-transformers==2.2.2
scikit-learn==1.3.0

# Basic PDF Processing
Pillow==10.0.1
tabula-py==2.9.0
easyocr==1.7.1
pytesseract==0.3.10

# Excel Processing and Visualization
matplotlib==3.8.2
seaborn==0.13.0
xlrd==2.0.1
xlsxwriter==3.1.9

# NLP and AI
transformers==4.35.2
torch==2.1.1
tokenizers==0.15.0

# Additional utilities
python-multipart==0.0.6
requests==2.31.0
urllib3==2.0.7

spacy==3.7.2